agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.1
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0
aws-opentelemetry-distro==0.10.0
boto3>=1.34.0
python-dotenv>=1.0.0
httpx>=0.24.0
uv>=0.1.0
PyJWT>=2.8.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0